Code structured. Comments deleted. Avatar animation repaired. Likes added. NavBar removed. Disign improved. Check email added(*)
