+++ Navigation/Navigation/PostViewController.swift
@@ -5,6 +5,8 @@ class PostViewController: UIViewController {
 
     var post: FeedViewController.Post?
     
+    
+//MARK: ==================================== INITs ====================================
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
@@ -18,6 +20,7 @@ class PostViewController: UIViewController {
         checkOrientation()
     }
 
+//MARK: ================================== ViewITEMs ==================================
     private func showBarButton() {
         let button = UIBarButtonItem(title: "Info...", style: .plain, target: self, action: #selector(tapAction))
         navigationItem.rightBarButtonItem = button
