--- Navigation/Navigation/Profile/PhotosCollectionViewCell.swift
@@ -1,44 +1,0 @@
-
-import UIKit
-
-class PhotosCollectionViewCell: UICollectionViewCell {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        //customizeCell()
-        showImageView()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private let imageView: UIImageView = {
-        $0.translatesAutoresizingMaskIntoConstraints = false
-        $0.contentMode = .scaleAspectFill
-        $0.clipsToBounds = true
-        return $0
-    }(UIImageView())
-    
-    func setupCell(_ photo: Photo) {
-        imageView.image = photo.imageName
-    }
-    
-    func showImageView() {
-        addSubview(imageView)
-        
-        NSLayoutConstraint.activate([
-            imageView.topAnchor.constraint(equalTo: topAnchor),
-            imageView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            imageView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            imageView.bottomAnchor.constraint(equalTo: bottomAnchor)
-        ])
-    }
-    
-    
-    
-//    private func customizeCell() {
-//        contentView.backgroundColor = .green
-//        contentView.layer.cornerRadius = 10
-//    }
-}
