+++ Navigation/Navigation/Profile/ProfileHeaderView.swift
@@ -1,41 +1,35 @@
 
 import UIKit
 
-public var screenWidth: CGFloat {
-    return UIScreen.main.bounds.width
-}
-public var screenHeight: CGFloat {
-    return UIScreen.main.bounds.height
-}
-
-public var bottomBlurView = NSLayoutConstraint()
-public var centerXProfileImage = NSLayoutConstraint()
-public var centerYProfileImage = NSLayoutConstraint()
-public var topProfileImage = NSLayoutConstraint()
-public var leadingProfileImage = NSLayoutConstraint()
-public var trailingProfileImage = NSLayoutConstraint()
-public var bottomProfileImage = NSLayoutConstraint()
-public var widthProfileImage = NSLayoutConstraint()
-public var heightProfileImage = NSLayoutConstraint()
-public let sizeProfileImage: CGFloat = 140
-
 class ProfileHeaderView: UIView {
-
-    private var statusText = "Waiting for something....."
     
     let space: CGFloat = 16
     
-    let headerView: UIView = {
-        let myView = UIView()
-        myView.translatesAutoresizingMaskIntoConstraints = false
-        myView.backgroundColor = .systemGray6
-        return myView
-    }()
+    
+    //MARK: ==================================== INITs ====================================
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        showProfileHeaderView()
+        tapGestureOnProfileImage()
+    }
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+    }
+    
+    
+    //MARK: ================================== ViewITEMs ==================================
+    
+    let viewUnderImage: UIView = {
+        $0.translatesAutoresizingMaskIntoConstraints = false
+        $0.backgroundColor = .clear
+        $0.alpha = 0.0
+        return $0
+    }(UIView())
     
     let profileImage: UIImageView = {
         let imageView = UIImageView()
         imageView.translatesAutoresizingMaskIntoConstraints = false
-        imageView.layer.cornerRadius = sizeProfileImage / 2          //размер 140 (140/2 - радиус)
+        imageView.layer.cornerRadius = sizeProfileImage / 2
         imageView.layer.borderWidth = 3
         imageView.layer.borderColor = UIColor.white.cgColor
         imageView.contentMode = .scaleAspectFill        //полное заполнение
@@ -49,7 +43,7 @@ class ProfileHeaderView: UIView {
     let buttonX: UIButton = {
         $0.translatesAutoresizingMaskIntoConstraints = false
         $0.layer.cornerRadius = 12
-        $0.backgroundColor = .systemGray
+        $0.backgroundColor = UIColor.AccentColor.normal
         $0.setImage(UIImage(systemName: "x.circle"), for: .normal)
         $0.tintColor = .white
         $0.alpha = 0.0
@@ -60,55 +54,31 @@ class ProfileHeaderView: UIView {
     
     @objc private func tapButtonX() {
         print("tap x")
-        bottomBlurView = blurView.bottomAnchor.constraint(equalTo: bottomAnchor)
-        blurView.setNeedsUpdateConstraints()
-        //headerView.addSubview(profileImage)
-        
-        UIView.animate(withDuration: 0.5, delay: 0.0, options: .curveEaseIn) { [self] in
-            centerXProfileImage = profileImage.centerXAnchor.constraint(equalTo: leadingAnchor, constant: space + sizeProfileImage / 2)
-            centerYProfileImage = profileImage.centerYAnchor.constraint(equalTo: topAnchor, constant: space + sizeProfileImage / 2)
-            widthProfileImage = profileImage.widthAnchor.constraint(equalToConstant: sizeProfileImage)
-            heightProfileImage = profileImage.heightAnchor.constraint(equalToConstant: sizeProfileImage)
-            NSLayoutConstraint.activate([centerXProfileImage, centerYProfileImage, widthProfileImage, heightProfileImage])
-
-            setNeedsUpdateConstraints()
-            blurView.setNeedsUpdateConstraints()
+        UIView.animate(withDuration: 1.5, delay: 0.0, options: .curveEaseIn) { [self] in
+            topProfileImage.constant = space
+            leadingProfileImage.constant = space
+            widthProfileImage.constant = sizeProfileImage
+            heightProfileImage.constant = sizeProfileImage
             profileImage.layer.cornerRadius = sizeProfileImage / 2
-            blurView.alpha = 0.0
-        } completion: { _ in   }
-
-        UIView.animate(withDuration: 0.3, delay: 0.0, options: .curveEaseOut) { [self] in
-            self.buttonX.alpha = 0.0
+            profileImage.layer.borderWidth = 3
+            viewUnderImage.alpha = 0.0
+            layoutIfNeeded()
         } completion: { _ in
-            self.buttonX.isHidden = true
+            UIView.animate(withDuration: 0.3, delay: 0.0, options: .curveEaseOut) { [self] in
+                buttonX.alpha = 0.0
+            } completion: { _ in
+                self.buttonX.isHidden = true
+            }
         }
     }
-    
-    func blurBackgroundEffect() -> UIVisualEffectView {
-        //код BlurEffect укороченный без констрейнтов
-        let effect = UIVisualEffectView(effect: UIBlurEffect(style: UIBlurEffect.Style.dark))
-        effect.alpha = 0.85
-        effect.translatesAutoresizingMaskIntoConstraints = false
-        effect.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-        //effect.frame = view.frame
-        return effect
-    }
-    
-    let blurView: UIView = {
-        $0.translatesAutoresizingMaskIntoConstraints = false
-        $0.backgroundColor = .none
-        $0.alpha = 0.0
-        return $0
-    }(UIView())
-    
+   
     
     //======================================================================================================
 
-    private let profileLabel: UITextView = {
-        let label = UITextView()
+    private let profileLabel: UILabel = {
+        let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
         label.font = UIFont.systemFont(ofSize: 18, weight: .bold)
-        label.isEditable = false
         label.text = "Obi Wan Kenobi"
         label.backgroundColor = UIColor(displayP3Red: 0.0, green: 0.0, blue: 0.0, alpha: 0.0)
         return label
@@ -122,9 +92,7 @@ class ProfileHeaderView: UIView {
         button.setTitle("Set status", for: .normal)
         button.setTitleColor(.white, for: .normal)
         button.setTitle("Status is being recorded", for: .highlighted)
-        //button.setTitleColor(.systemGray, for: .highlighted)
         button.addTarget(self, action: #selector(tapMainButton), for: .touchUpInside)
-
         button.layer.shadowOffset = CGSize(width: 4, height: 4)
         button.layer.shadowRadius = 4
         button.layer.shadowColor = UIColor.black.cgColor
@@ -147,18 +115,17 @@ class ProfileHeaderView: UIView {
         rotateAndSleep(0) //прячем вспомогательную кнопку без анимации
     }
     
-    private let profileStatus: UITextView = {
-        let status = UITextView()
+    let profileStatus: UILabel = {
+        let status = UILabel()
         status.translatesAutoresizingMaskIntoConstraints = false
         status.font = UIFont.systemFont(ofSize: 14, weight: .regular)
-        status.isEditable = false
         status.textColor = UIColor.darkGray
         status.text = "Waiting for something....."
         status.backgroundColor = UIColor(displayP3Red: 0.0, green: 0.0, blue: 0.0, alpha: 0.0)
         return status
     }()
 
-    private lazy var editStatus: UITextField = {
+    lazy var editStatus: UITextField = {
         let status = UITextField()
         status.font = UIFont.systemFont(ofSize: 15, weight: .regular)
         status.translatesAutoresizingMaskIntoConstraints = false
@@ -170,6 +137,7 @@ class ProfileHeaderView: UIView {
         status.layer.cornerRadius = 12
         status.layer.borderWidth = 1
         status.layer.borderColor = UIColor.black.cgColor
+        status.tintColor = UIColor.AccentColor.normal                          //цвет курсора
         status.layer.sublayerTransform = CATransform3DMakeTranslation(5, 0, 0) //сдвиг курсора на 5пт в textField (для красоты)
         status.addTarget(self, action: #selector(beginToEditStatus), for: .allEditingEvents)
         status.addTarget(self, action: #selector(changeStatusText), for: .editingChanged)
@@ -191,7 +159,7 @@ class ProfileHeaderView: UIView {
     }
     
     private lazy var statusAlert: UILabel = {
-        //$0.translatesAutoresizingMaskIntoConstraints = false
+        $0.translatesAutoresizingMaskIntoConstraints = false
         $0.text = ""
         $0.textColor = .systemRed
         $0.textAlignment = .right
@@ -218,6 +186,62 @@ class ProfileHeaderView: UIView {
         editStatus.text = ""
         rotateAndSleep(1) //прячем вспомогательную кнопку с анимацией (после нажатия на неё)
     }
+    
+    func showProfileHeaderView() {
+        [profileLabel, mainButton, profileStatus, editStatus, buttonAccept].forEach { self.addSubview($0) }
+        
+        NSLayoutConstraint.activate([
+            profileLabel.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: 27),
+            profileLabel.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor, constant: sizeProfileImage + 2 * space),
+            profileLabel.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: -space),
+            profileLabel.heightAnchor.constraint(equalToConstant: 30),
+
+            mainButton.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: sizeProfileImage + 2 * space),
+            mainButton.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor, constant: space),
+            mainButton.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: -space),
+            mainButton.heightAnchor.constraint(equalToConstant: 50),
+
+            profileStatus.bottomAnchor.constraint(equalTo: mainButton.topAnchor, constant: -space - 46),
+            profileStatus.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor, constant: sizeProfileImage + 2 * space),
+            profileStatus.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: -space),
+            profileStatus.heightAnchor.constraint(equalToConstant: 26),
+
+            editStatus.bottomAnchor.constraint(equalTo: mainButton.topAnchor, constant: -space / 2 - 6),
+            editStatus.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor, constant: sizeProfileImage + 2 * space),
+            editStatus.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: -space),
+            editStatus.heightAnchor.constraint(equalToConstant: 40),
+
+            buttonAccept.topAnchor.constraint(equalTo: editStatus.topAnchor, constant: 5),
+            buttonAccept.trailingAnchor.constraint(equalTo: editStatus.trailingAnchor, constant: -5),
+            buttonAccept.widthAnchor.constraint(equalToConstant: 30),
+            buttonAccept.heightAnchor.constraint(equalToConstant: 30)
+        ])
+       
+        [viewUnderImage, profileImage, buttonX].forEach { self.addSubview($0) }
+        viewUnderImage.addSubview(blurBackgroundEffect())
+               
+        topProfileImage = profileImage.topAnchor.constraint(equalTo: viewUnderImage.topAnchor, constant: space)
+        leadingProfileImage = profileImage.leadingAnchor.constraint(equalTo: viewUnderImage.leadingAnchor, constant: space)
+        widthProfileImage = profileImage.widthAnchor.constraint(equalToConstant: sizeProfileImage)
+        heightProfileImage = profileImage.heightAnchor.constraint(equalToConstant: sizeProfileImage)
+        
+        NSLayoutConstraint.activate([
+            viewUnderImage.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor),
+            viewUnderImage.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor),
+            viewUnderImage.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor),
+            viewUnderImage.heightAnchor.constraint(equalToConstant: screenHeight),
+            
+            topProfileImage, leadingProfileImage, widthProfileImage, heightProfileImage,
+            
+            buttonX.topAnchor.constraint(equalTo: viewUnderImage.topAnchor, constant: 20),
+            buttonX.trailingAnchor.constraint(equalTo: viewUnderImage.trailingAnchor, constant: -20),
+            buttonX.widthAnchor.constraint(equalToConstant: 24),
+            buttonX.heightAnchor.constraint(equalToConstant: 24)
+        ])
+    }
+
+    
+//MARK: =================================== METHODs ===================================
     private func rotateAndSleep(_ key: Int) {
         if key == 1 { //анимация вращения вокруг своей оси на вспомогательной кнопке
             var perspective = CATransform3DIdentity
@@ -235,81 +259,57 @@ class ProfileHeaderView: UIView {
         })
     }
     
-    func showProfileHeaderView() {
-        addSubview(headerView)
-        
-        NSLayoutConstraint.activate([
-            headerView.topAnchor.constraint(equalTo: topAnchor),
-            headerView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            headerView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            headerView.bottomAnchor.constraint(equalTo: bottomAnchor),
-            headerView.heightAnchor.constraint(equalToConstant: 240)
-        ])
-        
-        [profileLabel, mainButton, profileStatus, editStatus, buttonAccept].forEach { headerView.addSubview($0) }
-        
-        NSLayoutConstraint.activate([
-            profileLabel.topAnchor.constraint(equalTo: headerView.topAnchor, constant: 27),
-            profileLabel.leadingAnchor.constraint(equalTo: headerView.leadingAnchor, constant: sizeProfileImage + 2 * space),
-            profileLabel.trailingAnchor.constraint(equalTo: headerView.trailingAnchor, constant: -space),
-            profileLabel.heightAnchor.constraint(equalToConstant: 30),
+    public var centerXProfileImage = NSLayoutConstraint()
+    public var centerYProfileImage = NSLayoutConstraint()
+    public var topProfileImage = NSLayoutConstraint()
+    public var leadingProfileImage = NSLayoutConstraint()
+    public var trailingProfileImage = NSLayoutConstraint()
+    public var bottomProfileImage = NSLayoutConstraint()
+    public var widthProfileImage = NSLayoutConstraint()
+    public var heightProfileImage = NSLayoutConstraint()
 
-            mainButton.topAnchor.constraint(equalTo: headerView.topAnchor, constant: sizeProfileImage + 2 * space),
-            mainButton.leadingAnchor.constraint(equalTo: headerView.leadingAnchor, constant: space),
-            mainButton.trailingAnchor.constraint(equalTo: headerView.trailingAnchor, constant: -space),
-            mainButton.heightAnchor.constraint(equalToConstant: 50),
-
-            profileStatus.bottomAnchor.constraint(equalTo: mainButton.topAnchor, constant: -space - 40),
-            profileStatus.leadingAnchor.constraint(equalTo: headerView.leadingAnchor, constant: sizeProfileImage + 2 * space),
-            profileStatus.trailingAnchor.constraint(equalTo: headerView.trailingAnchor, constant: -space),
-            profileStatus.heightAnchor.constraint(equalToConstant: 26),
-
-            editStatus.bottomAnchor.constraint(equalTo: mainButton.topAnchor, constant: -space / 2),
-            editStatus.leadingAnchor.constraint(equalTo: headerView.leadingAnchor, constant: sizeProfileImage + 2 * space),
-            editStatus.trailingAnchor.constraint(equalTo: headerView.trailingAnchor, constant: -space),
-            editStatus.heightAnchor.constraint(equalToConstant: 40),
-
-            buttonAccept.topAnchor.constraint(equalTo: editStatus.topAnchor, constant: 5),
-            buttonAccept.trailingAnchor.constraint(equalTo: editStatus.trailingAnchor, constant: -5),
-            buttonAccept.widthAnchor.constraint(equalToConstant: 30),
-            buttonAccept.heightAnchor.constraint(equalToConstant: 30)
-        ])
-        
-        [blurView, profileImage, buttonX].forEach { headerView.addSubview($0) }
-        blurView.addSubview(blurBackgroundEffect())
-        
-        bottomBlurView = blurView.bottomAnchor.constraint(equalTo: bottomAnchor)
-        //topProfileImage = profileImage.topAnchor.constraint(equalTo: blurView.topAnchor, constant: space)
-        //leadingProfileImage = profileImage.leadingAnchor.constraint(equalTo: blurView.leadingAnchor, constant: space)
-        centerXProfileImage = profileImage.centerXAnchor.constraint(equalTo: headerView.leadingAnchor, constant: space + sizeProfileImage / 2)
-        centerYProfileImage = profileImage.centerYAnchor.constraint(equalTo: headerView.topAnchor, constant: space + sizeProfileImage / 2)
-        widthProfileImage = profileImage.widthAnchor.constraint(equalToConstant: sizeProfileImage)
-        heightProfileImage = profileImage.heightAnchor.constraint(equalToConstant: sizeProfileImage)
-        
-        NSLayoutConstraint.activate([
-            blurView.topAnchor.constraint(equalTo: topAnchor),
-            blurView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            blurView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            bottomBlurView,
-            
-            /*topProfileImage, leadingProfileImage,*/
-            centerXProfileImage, centerYProfileImage, widthProfileImage, heightProfileImage,
-            
-            buttonX.topAnchor.constraint(equalTo: blurView.topAnchor, constant: 20),
-            buttonX.trailingAnchor.constraint(equalTo: blurView.trailingAnchor, constant: -20),
-            buttonX.widthAnchor.constraint(equalToConstant: 24),
-            buttonX.heightAnchor.constraint(equalToConstant: 24)
-        ])
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        showProfileHeaderView()
-    }
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-    }
     
     //MARK: жесты и анимация
+    func tapGestureOnProfileImage() {
+        print("tapGesture?")
+        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(tapOnImage))
+        profileImage.addGestureRecognizer(tapGesture)
+    }
 
+    @objc private func tapOnImage() {
+        print("tap")
+        UIView.animate(withDuration: 2.5, delay: 0.0, options: .curveEaseOut) { [self] in
+            viewUnderImage.addSubview(blurBackgroundEffect())
+            if (UIScreen.main.bounds.height > UIScreen.main.bounds.width) {
+                topProfileImage.constant = UIScreen.main.bounds.height / 2 - absoluteWidth / 2 - 80
+                leadingProfileImage.constant = 0
+                trailingProfileImage.constant = 0
+            } else {
+                topProfileImage.constant = 0
+                leadingProfileImage.constant = UIScreen.main.bounds.width / 2 - absoluteWidth / 2
+                trailingProfileImage.constant = UIScreen.main.bounds.width / 2 + absoluteWidth / 2
+            }
+            widthProfileImage.constant = absoluteWidth
+            heightProfileImage.constant = absoluteWidth
+            profileImage.layer.cornerRadius = 0
+            profileImage.layer.borderWidth = 0
+            viewUnderImage.alpha = 0.85
+            layoutIfNeeded()
+            viewUnderImage.layoutIfNeeded()
+        } completion: { _ in
+            UIView.animate(withDuration: 0.3, delay: 0.0, options: .curveEaseOut) { [self] in
+                buttonX.isHidden = false
+                buttonX.alpha = 1.0
+            } completion: { _ in  }
+        }
+    }
+    
+    func blurBackgroundEffect() -> UIVisualEffectView {
+        //код BlurEffect укороченный без констрейнтов
+        let effect = UIVisualEffectView(effect: UIBlurEffect(style: UIBlurEffect.Style.dark))
+        effect.alpha = 0.85
+        effect.translatesAutoresizingMaskIntoConstraints = false
+        effect.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        return effect
+    }
 }
