+++ Navigation/Navigation/Profile/ProfileTableViewCell.swift
@@ -3,6 +3,24 @@ import UIKit
 
 class ProfileTableViewCell: UITableViewCell {
 
+    var index: IndexPath?
+    var post: Post?
+    weak var delegate: AddLikeDelegate?
+    
+    
+//MARK: ==================================== INITs ====================================
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        show()
+        setupLikesGestures()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    
+//MARK: ================================== ViewITEMs ==================================
     private let postView: UIView = {
         $0.translatesAutoresizingMaskIntoConstraints = false
         $0.backgroundColor = .white
@@ -11,21 +29,16 @@ class ProfileTableViewCell: UITableViewCell {
     
     let postName: UILabel = {
         $0.translatesAutoresizingMaskIntoConstraints = false
-        //$0.backgroundColor = .lightGray
         $0.font = UIFont.systemFont(ofSize: 20, weight: .bold)
-        //$0.text = "Мама мыла раму, Вова курит Приму."
         $0.numberOfLines = 2
         return $0
     }(UILabel())
     
     let postDescription: UITextView = {
         $0.translatesAutoresizingMaskIntoConstraints = false
-        //$0.backgroundColor = .white
         $0.font = UIFont.systemFont(ofSize: 14, weight: .regular)
         $0.textColor = .systemGray
         $0.isEditable = false
-        //$0.text = "Быть или не быть - вот в чём вопрос! Во поле березка стояла, во поле кудрявая стояла. Люли-люли - стояла! Нас орда! А нас рать! Однажды в студёную зимнюю пору, я из лесу вышел - был сильный мороз. Смотрю поднимается медленно в гору, лошадка везущая хворосту воз. Откуда дровишки - Из лесу, вестимо. Отец, слышишь, рубит, а я отвожу."
-        //$0.numberOfLines = 2
         $0.isScrollEnabled = false
         return $0
     }(UITextView())
@@ -41,10 +54,8 @@ class ProfileTableViewCell: UITableViewCell {
         $0.translatesAutoresizingMaskIntoConstraints = false
         $0.font = UIFont.systemFont(ofSize: 16, weight: .regular)
         $0.textColor = .black
-        //$0.backgroundColor = .lightGray
-        //$0.text = "Likes: 111"
         $0.textAlignment = .left
-        //$0.nu
+        $0.isUserInteractionEnabled = true
         return $0
     }(UILabel())
     
@@ -52,35 +63,10 @@ class ProfileTableViewCell: UITableViewCell {
         $0.translatesAutoresizingMaskIntoConstraints = false
         $0.font = UIFont.systemFont(ofSize: 16, weight: .regular)
         $0.textColor = .black
-        //$0.backgroundColor = .lightGray
-        //$0.text = "Viewes: 555"
         $0.textAlignment = .right
         return $0
     }(UILabel())
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        show()
-        //customizecell()
-    }
     
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func setupCell(_ post: Post) {
-        postImage.image = post.imageName
-        postName.text = post.author
-        postDescription.text = post.description
-        likes.text = "Likes: " + String(post.likes)
-        views.text = "Views: " + String(post.views)
-    }
-    
-    //func customizecell() {
-        //postView.layer.cornerRadius = 10
-        //postView.layer.borderWidth = 10
-        //postView.layer.borderColor = UIColor.purple.cgColor
-    //}
     func show() {
         [postView, postName, postDescription, postImage, likes, views].forEach { contentView.addSubview($0) }
         
@@ -93,7 +79,6 @@ class ProfileTableViewCell: UITableViewCell {
             postName.topAnchor.constraint(equalTo: postView.topAnchor, constant: 16),
             postName.leadingAnchor.constraint(equalTo: postView.leadingAnchor, constant: 16),
             postName.trailingAnchor.constraint(equalTo: postView.trailingAnchor, constant: -16),
-            //postName.heightAnchor.constraint(equalToConstant: 20)
 
             postImage.topAnchor.constraint(equalTo: postName.bottomAnchor, constant: 12),
             postImage.leadingAnchor.constraint(equalTo: postView.leadingAnchor),
@@ -103,18 +88,35 @@ class ProfileTableViewCell: UITableViewCell {
             postDescription.topAnchor.constraint(equalTo: postImage.bottomAnchor, constant: 16),
             postDescription.leadingAnchor.constraint(equalTo: postName.leadingAnchor),
             postDescription.trailingAnchor.constraint(equalTo: postName.trailingAnchor),
-            //postDescription.heightAnchor.constraint(equalToConstant: )
 
             likes.topAnchor.constraint(equalTo: postDescription.bottomAnchor, constant: 16),
             likes.leadingAnchor.constraint(equalTo: postView.leadingAnchor, constant: 16),
-            likes.trailingAnchor.constraint(equalTo: postView.trailingAnchor, constant: -16),
+            likes.trailingAnchor.constraint(equalTo: postView.centerXAnchor, constant: 0),
             likes.bottomAnchor.constraint(equalTo: postView.bottomAnchor, constant: -16),
             
             views.topAnchor.constraint(equalTo: postDescription.bottomAnchor, constant: 16),
-            views.leadingAnchor.constraint(equalTo: postView.leadingAnchor, constant: 16),
+            views.leadingAnchor.constraint(equalTo: postView.centerXAnchor, constant: 0),
             views.trailingAnchor.constraint(equalTo: postView.trailingAnchor, constant: -16),
             views.bottomAnchor.constraint(equalTo: postView.bottomAnchor, constant: -16)
         ])
     }
     
+  
+//MARK: =================================== METHODs ===================================
+    func setupCell(_ post: Post) {
+        postImage.image = post.imageName
+        postName.text = post.author
+        postDescription.text = post.description
+        likes.text = "Likes: \(post.likes)"
+        views.text = "Views: \(post.views)"
+    }
+
+    func setupLikesGestures() {
+        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(tapLike))
+        likes.addGestureRecognizer(tapGesture)
+    }
+    @objc private func tapLike (){
+        if let index = index { delegate?.addLike(index, "Profile") }
+        if let post = post { likes.text = "Likes: \(post.likes)" }
+    }
 }
