--- Navigation/Navigation/Profile/ShowViewController.swift
@@ -1,17 +1,0 @@
-
-import UIKit
-
-class ShowViewController: UIViewController {
-
-    let profileTableViewCell = ProfileTableViewCell()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        profileTableViewCell.show()
-    }
-    
-    
-    
-    
-
-}
