--- Navigation/Navigation/Profile/UIView+Extension.swift
@@ -1,8 +1,0 @@
-
-import UIKit
-
-extension UIView {
-    static var identifier: String {
-        return String(describing: self)
-    }
-}
