+++ Navigation/Navigation/Related data/UIView+Extension.swift
@@ -1,0 +1,8 @@
+
+import UIKit
+
+extension UIView {
+    static var identifier: String {
+        return String(describing: self)
+    }
+}
